Requested changes updated: (1) changed folder name 'heatmaps' to 'heatmap' (2) removed anonymous function for getting legend values (3) changed test cases code structure to using only one loop (4) removed 'seleniumAddress' in 'protractor.conf.js'
