+++ test/e2e/specs/heatmap/heatmap-expected-value.js
