+++ test/e2e/specs/protractor.conf.js
@@ -2,7 +2,6 @@ let SpecReporter = require('jasmine-spec-reporter').SpecReporter;
 
 exports.config = {
     framework : 'jasmine',
-    seleniumAddress : 'http://localhost:4444/wd/hub',
     specs : ['*.spec.js'],
     capabilities : {
         browserName : 'chrome'
